requirements.txt
Django==2.0.6
django-ckeditor==5.5.0
django-js-asset==1.1.0